All user and community routes now working. added support for update users and communities. added support for authorization, so only authorized users can update or delete users or communities. Updated requirements.txt

+++ server/routers/user_router.py
@@ -1,4 +1,4 @@
-from fastapi import Depends, HTTPException, APIRouter
+from fastapi import Depends, HTTPException, APIRouter, Request
 from sqlalchemy.orm import Session
 
 from services.database import get_db
@@ -30,21 +30,21 @@ def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
 @router.post("/", response_model=schemas.User)
 def create_user(
     user: schemas.UserCreate,
-    current_user: dict = Depends(authentication.authenticate_user),
+    current_user: str = Depends(authentication.authenticate_user),
     db: Session = Depends(get_db),
 ):
     db_user = crud.get_user_by_email(db, email=user.email)
     if db_user:
         raise HTTPException(status_code=400, detail="Email already registered")
 
-    db_user = crud.get_user(db, user_id=current_user["uid"])
+    db_user = crud.get_user(db, user_id=current_user)
     if db_user:
         raise HTTPException(
             status_code=400, detail="An user with this ID already exists"
         )
 
     user_data = {
-        "firebase_uid": current_user["uid"],
+        "firebase_uid": current_user,
         "email": user.email,
         "first_name": user.first_name,
         "last_name": user.last_name,
@@ -53,18 +53,41 @@ def create_user(
     return crud.create_user(db=db, user=user_data)
 
 
-# UPDATE USER STILL NEEDS IMPLEMENTATION
+# UPDATE USER
 @router.put("/{user_id}", response_model=schemas.User)
-def update_user(user_id: str, db: Session = Depends(get_db)):
-    updated_user = crud.update_user(db, user_id=user_id)
-    if update_user is None:
+def update_user(
+    user: schemas.UserUpdate,
+    user_id: str,
+    request_user_id: str = Depends(authentication.authenticate_user),
+    db: Session = Depends(get_db),
+):
+    db_user = crud.get_user(db, user_id=user_id)
+    if db_user is None:
         raise HTTPException(status_code=404, detail="User not found")
+    if user_id != request_user_id:
+        raise HTTPException(
+            status_code=403, detail="Not authorized to update this user"
+        )
+
+    updated_user = crud.update_user(
+        db=db, user=db_user, update_data=user.dict(exclude_unset=True)
+    )
+    if updated_user is None:
+        raise HTTPException(status_code=400, detail="User could not be updated")
     return updated_user
 
 
 # DELETE USER
 @router.delete("/{user_id}", response_model=schemas.User)
-def delete_user(user_id: str, db: Session = Depends(get_db)):
+def delete_user(
+    user_id: str,
+    request_user_id: str = Depends(authentication.authenticate_user),
+    db: Session = Depends(get_db),
+):
+    if user_id != request_user_id:
+        raise HTTPException(
+            status_code=403, detail="Not authorized to delete this user"
+        )
     deleted_user = crud.delete_user(db, user_id=user_id)
     if not deleted_user:
         raise HTTPException(status_code=404, detail="User not found")
